refactor(root): Modular hexagonal architecture
