+++ backend-express/interface/controllers/user.ts
@@ -1,13 +1,16 @@
 import { NextFunction, Request, Response } from "express";
 import jwt from 'jsonwebtoken';
 import bcrypt from 'bcryptjs';
+import dotenv from "dotenv"
+
 
 import {ReadAll, ReadByEmail, ReadById} from "../../../core/application/usecases/atomic/user"
 import {FindDbError, InvalidUrlError, SetEnvError, UnauthorizedError} from "../../../core/domain/errors/main"
-import { PrismaUserRepository } from "../../infrastructure/repositories/prisma-user";
-import { User } from "@prisma/client";
+import { PrismaUserRepository } from "../../../infrastructure/repositories/prisma-user";
+import { User } from "../../../core/domain/entities/User";
 export const userRepository = new PrismaUserRepository()
 
+dotenv.config()
 export class UserController {
     constructor(){
         this.read = this.read.bind(this);
