+++ backend-express/package.json
@@ -11,7 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^5.19.1",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
@@ -28,7 +27,6 @@
     "@types/node": "^22.5.5",
     "esbuild-register": "^3.6.0",
     "nodemon": "^3.1.7",
-    "prisma": "^5.19.1",
     "typescript": "^5.6.2"
   }
 }
