+++ backend-express/requests.rest
@@ -13,14 +13,14 @@ Content-Type: application/json
 ### Read todos los usuarios ❌ 
 GET http://localhost:3000/admins/users
 Accept: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6IkFETUlOIiwiaWF0IjoxNzI3MDg2ODI3LCJleHAiOjE3MjcwOTA0Mjd9.6Chcui95aGHee_W7Gut7dwk1GkPt-8CKZm---4NczEM
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Miwicm9sZSI6IlVTRVIiLCJpYXQiOjE3MjcxMDI5NDEsImV4cCI6MTcyNzEwNjU0MX0.cgUWwF0l5z16DlKle2bfdr5dSL35SUed95dAWLs8DNY
 
 ###
 
 ### Read un usuario por ID ✅
 GET http://localhost:3000/users/id?q=2
 Accept: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Miwicm9sZSI6IlVTRVIiLCJpYXQiOjE3MjcwOTc0NzgsImV4cCI6MTcyNzEwMTA3OH0.qVS-2ZI5BxyqCNE2rPrz_qwfN61tpXPjSr1edbQGcas
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Miwicm9sZSI6IlVTRVIiLCJpYXQiOjE3MjcxMDI5NDEsImV4cCI6MTcyNzEwNjU0MX0.cgUWwF0l5z16DlKle2bfdr5dSL35SUed95dAWLs8DNY
 
 ###
