+++ core/tsconfig.json
@@ -1,13 +1,17 @@
 {
     "compilerOptions": {
       "baseUrl": ".",
-      "module": "ESNext",
+      "module": "CommonJS",
       "target": "ES2020",
-      "lib": ["dom", "esnext"],
-      "esModuleInterop": true,
-      "moduleResolution": "node",
-      "resolveJsonModule": true,
-      "skipLibCheck": true
+      // "rootDir": "../", // ❓ Establece la carpeta raíz, sino la raíz es donde esta tsconfig.json
+      "lib": [ "esnext"],
+      // "lib": ["dom", "esnext"], // ✅⚠️ Biblotecas de tipos que importara ❓
+      // "moduleResolution": "node", // ✅ Resolucion de modulos
+      // "resolveJsonModule": true, // ✅ Permite importar archivos json 
+      "esModuleInterop": true, // ✅ Interoperabilidad entre CommonJS y ES6
+      "skipLibCheck": true, // ❓ Omitir verificación de tipos en .d.ts
+      "strict": true, 
+      "forceConsistentCasingInFileNames": true // ✅ Forzar mayúsculas y minúsculas en nombres de archivo
     },
     "include": ["./domain", "./application"]
   }
