+++ infrastructure/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "infrastructure",
+  "version": "1.0.0",
+  "scripts": {
+    
+  },
+  "dependencies": {
+  "@prisma/client": "^5.19.1"
+  },
+  "devDependencies": {
+    "prisma": "^5.19.1"
+  }
+}
