+++ infrastructure/repositories/prisma-likepost.ts
@@ -1,5 +1,5 @@
 import { PrismaClientConfig } from "../connectors/prisma-db";
-import { LikePostRepository } from "../../../core/application/repositories/likepost";
+import { LikePostRepository } from "../../core/application/repositories/likepost";
 
 export class PrismaLikePostRepository extends PrismaClientConfig implements LikePostRepository {
     async create(postId: number, userId: number): Promise<Object> {
