+++ infrastructure/repositories/prisma-user.ts
@@ -1,8 +1,8 @@
 // backend/src/infrastructure/repositories/PrismaUserRepository.ts
 
 import { Prisma } from '@prisma/client';
-import {UserRepository}from "../../../core/application/repositories/user"
-import {User} from "../../../core/domain/entities/User"
+import {UserRepository}from "../../core/application/repositories/user"
+import {User} from "../../core/domain/entities/User"
 import { PrismaClientConfig } from '../connectors/prisma-db';
 
 export class PrismaUserRepository extends PrismaClientConfig implements UserRepository {
