+++ infrastructure/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "CommonJS",
+    // "rootDir": "../", 
+    "lib": [ "esnext"],
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true
+  },
+  "include": ["./prisma", "./repositories", "./connectors"],
+  "exclude": ["node_modules"]
+}
